packages/backend/requirements.txt
# requirements.txt
python-dotenv
requests
google-generativeai
serpapi
fastapi
uvicorn[standard]
google-adk
weasyprint
markdown-it-py
sqlalchemy       
passlib[bcrypt]    
python-jose[cryptography] 
multipart          